rename: loadingspinner.tsx 파일 공통 컴포넌트 폴더로 위치 변경 및 위치 변경으로 인한 관련 import url 수정
