+++ src/mock/msw/browser.ts
@@ -1,5 +1,5 @@
 import { setupWorker } from 'msw/browser';
-import { handlers } from './handlers';
+import { handlers } from './projectHandlers';
 import { dashboardHandlers } from './dashboardHandler';
 
 export const worker = setupWorker(...handlers, ...dashboardHandlers);
